type: refactor
This overhauls the detailPost in state to be a copy of the object rather than an ID.  This works well enough with the current array setup but could be improved.  This removes any separate state data from the PostDetail component.  This commit also fixes the previous bug of correctly updating state to reflect Post edits on submission.