type: feature
Implements voting mechanism on Post, PostDetail, and Comment components. This has no validation or user restriction yet and could be expanded in future commits. This adds a voting action with path to be set for 'posts' or 'comments', the id, and whether it is an 'upVote' or 'downVote'. Similar arguments are in place for the API and the reducers determine what to do based on the provided path.